+++ yumyum_front/src/views/MyPage/MypageUpdate.tsx
@@ -1,206 +1,248 @@
-import { css } from "@emotion/react"
+/** @jsxImportSource @emotion/react */
+import axios from "axios";
+import React, { useEffect, useState } from "react";
+import * as s from "./MypageCss";
+import { Link, useNavigate } from "react-router-dom";
+import { HOME_PATH, MAIN_PATH, MY_PAGE, MY_PAGE_UPDATE } from "../../constants";
 
-export const all = css`
-  box-sizing: border-box;
-  padding: 20px;
-  margin-bottom: 100px;
-  width: 100%;
-  height: 100%;
-`
+interface User {
+  userId: string;
+  userPw: string;
+  userConfirmPw: string;
+  userName: string;
+  userEmail: string;
+  userPhone: string;
+  userBusinessNumber: string;
+  marketingAgreed: boolean;
+}
 
-export const header = css`
-  box-sizing: border-box;
-  display: flex;
-  margin-bottom: 20px;
-  padding: 0px 30px;
-  width: 100%;
-  height: 50px;
-  font-size: 35px;
+export default function Mypage() {
+  const [user, setUser] = useState<User>({
+    userId: "",
+    userPw: "",
+    userConfirmPw: "",
+    userName: "",
+    userEmail: "",
+    userPhone: "",
+    userBusinessNumber: "",
+    marketingAgreed: false,
+  });
 
-`
+  const [id, SetId] = useState<Number>(1);
+  const [errorMsg, setErrorMsg] = useState<{ [key: string]: string }>({});
+  const [successMsg, setSuccessMsg] = useState<{ [key: string]: string }>({});
 
-export const body = css`
-  box-sizing: border-box;
-  margin-bottom: 20px;
-  border: 1px solid #dbdbdb;
-  width: 90%;
-  height: 650px;
-  background-color: white;
-  border-radius: 10px;
-  
-`
+  const navigate = useNavigate()
 
-export const updateBody = css`
-  box-sizing: border-box;
-  margin-bottom: 20px;
-  border: 1px solid #dbdbdb;
-  width: 85% ;
-  height: 850px;
-  background-color: white;
-  border-radius: 10px;
-  
-`
+  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = e.target;
+    let errorMsg = "";
 
-export const user = css`
-  box-sizing: border-box;
-  display: flex;
-  margin: 5px 10px;
-  border: 1px solid #dbdbdb;
-  border-radius: 5px;
-  width: 50%;
-  height: 50px;
-  font-size: 19px;
-  justify-content: left;
-  
-  
-`
+    switch (name) {
+      case "userPw":
+        const passwordRegex =
+          /^(?=.*[A-Za-z])(?=.*\d)(?=.*[!@#$%^&*])[^\s]{8,15}$/;
+        if (!passwordRegex.test(value)) {
+          errorMsg =
+            "비밀번호는 영문자, 숫자, 특수문자를 포함하여 8~15자 이내로 입력해주세요.";
+        }
+        break;
+      case "userConfirmPw":
+        if (value !== user.userPw) {
+          errorMsg = "비밀번호가 일치하지 않습니다.";
+        } else {
+          setSuccessMsg((prev) => ({
+            ...prev,
+            userConfirmPw: "비밀번호가 일치합니다.",
+          }));
+          
+        }
+        break;
+      case "userPhone":
+        const phoneRegex = /^\d{10,11}$/;
+        if (!phoneRegex.test(value)) {
+          errorMsg = "유효한 휴대폰 번호를 입력해주세요. (10~11자리)";
+        }
+        break;
+      default:
+        break;
+    }
 
-export const userDetail = css`
-  border: none;
-  font-size: 19px;
-  width: 100%;
-  padding-left: 20px;
-  cursor: default;
-  &:read-only {
-    background-color: #f0f0f0;
-    color: #333;
-  }
-`
+    setErrorMsg((prev) => ({
+      ...prev,
+      [name]: errorMsg,
+    }));
 
-export const checkBoxInput = css`
-  box-sizing: border-box;
-  display: flex;
-  width: 20px;
-  height: 20px;
-  
-`
+    setUser((prevUser) => ({
+      ...prevUser,
+      [name]: value,
+    }));
+  };
 
-export const sort = css`
-  display: flex;
-  flex-wrap: wrap;
-  margin: 15px;
-  padding: 10px;
-  align-items: flex-start;
-  & > div:nth-of-type(1) {
-    width: 200px;
-  }
-`
+  const handleCheckBox = () => {
+    setUser((prevUser) => ({
+      ...prevUser,
+      marketingAgreed: !prevUser.marketingAgreed,
+    }));
+  };
 
-export const myProfile = css`
-  display: flex;
-  margin-left: 50px;
-  align-items: center;
-  margin-top: 22px;
-  font-size: 16px;
+  const fetchData = async () => {
+    try {
+      const userData = await axios.get(
+        `http://localhost:4041/api/v1/mypage/${id}`
+      );
+      setUser(userData.data.data);
+    } catch (e) {
+      console.error("데이터를 불러오지 못했습니다.", e);
+    }
+  };
+  useEffect(() => {
+    fetchData();
+  }, []);
 
-`
-export const myProfile_market = css`
-  display: flex;
-  margin-left: 50px;
-  align-items: center;
-  margin-top: 5px;
-  font-size: 16px;
-
-`
-export const myProfile_market_div = css`
-  padding-top: 5px;
-  font-size: 16px;
-
-`
-
-export const dialogContent = css`
-  background-color: white; 
-  padding: 20px; 
-  border-radius: 5px; 
-  text-align:center
-`
-
-export const buttonContainer = css`
-  display: flex;
-  justify-content: center;
-  text-align: center;
-`
-
-export const updateUser = css`
-  border: 1px solid #808;
-  border-radius: 3px;
-  color: #808;
-  background-color: #fff;
-  margin: 0px 10px;
-  padding: 10px 30px;
-  display: inline-block;
-  text-align: center;
-  text-decoration: none;
-  font-size: 13px;
-  text-align: center;
-  line-height: 17.778px;
-  cursor: pointer;
-
-  &:hover{
-    background-color: #f0f0f0;
+  const handlePutUser = async (e: React.FormEvent) => {
+    e.preventDefault();
+    const hasErrors = Object.entries(errorMsg).some(([key, msg]) => key !== "form" && msg !== "")
+    if (hasErrors) {
+      setErrorMsg((prev) => ({ ...prev, form: "Errors in errorText "}));
+      return
+    }
+    try{
+      await axios.put(`http://localhost:4041/api/v1/mypage/update/${id}`, user);
+      console.log(user);
+      alert("수정이 완료되었습니다.")
+      navigate(MAIN_PATH)
+    } catch(e) {
+      console.error("데이터를 불러오지 못했습니다.", e);
+    }
   }
 
-  a {
-    color: #808;
-    text-decoration: none;
-    display: block;
-  }
-  
-`
+  return (
+    <div css={s.all}>
+      <div css={s.header}>
+        <h1>마이페이지</h1>
+      </div>
+      <hr style={{ border: "2px solid #000", marginTop: "0px", marginBottom: "10px" }} />
+      <div css={s.updateBody}>
+        <div css={s.sort}>
+          <div css={s.myProfile}>아이디</div>
+          <div css={s.user}>
+            <input
+              type="text"
+              css={s.updateUserDetail}
+              value={user.userId}
+              readOnly
+            />
+          </div>
+        </div>
+        <div css={s.sort}>
+          <div css={s.myProfile}>비밀번호</div>
+          <div css={s.user}>
+            <input
+              type="password"
+              name="userPw"
+              css={s.updateUserDetail}
+              placeholder="비밀번호를 입력하세요"
+              onChange={handleInputChange}
+            />
+          </div>
+          {errorMsg.userPw && <div css={s.errorMsg}>{errorMsg.userPw}</div>}
+        </div>
+        <div css={s.sort}>
+          <div css={s.myProfile}>비밀번호 확인</div>
+          <div css={s.user}>
+              <input
+                type="password"
+                name="userConfirmPw"
+                css={s.updateUserDetail}
+                placeholder="다시 비밀번호를 입력하세요"
+                onChange={handleInputChange}
+              />            
+          </div>
+          {user.userPw === user.userConfirmPw ? (
+            <div css={s.successMsg}>{successMsg.userConfirmPw}</div>
+          ) : errorMsg.userConfirmPw && (
+            <div css={s.errorMsg}>{errorMsg.userConfirmPw}</div>
+          )}
+        </div>
 
-export const deleteUser = css`
-border: 1px solid #808;
-  border-radius: 3px;
-  color: #808;
-  background-color: #fff;
-  margin: 0px 10px;
-  padding: 10px 30px;
-  display: inline-block;
-  text-align: center;
-  cursor: pointer;
+        <div css={s.sort}>
+          <div css={s.myProfile}>이름</div>
+          <div css={s.user}>
+            <input
+              type="text"
+              css={s.updateUserDetail}
+              value={user.userName}
+              readOnly
+            />
+          </div>
+        </div>
+        <div css={s.sort}>
+          <div css={s.myProfile}>이메일</div>
+          <div css={s.user}>
+            <input
+              type="text"
+              css={s.updateUserDetail}
+              value={user.userEmail}
+              readOnly
+            />
+          </div>
+        </div>
+        <div css={s.sort}>
+          <div css={s.myProfile}>핸드폰 번호</div>
+          <div css={s.user}>
+            <input
+              type="text"
+              name="userPhone"
+              css={s.updateUserDetail}
+              value={user.userPhone}
+              onChange={handleInputChange}
+            />
+          </div>
+          {errorMsg.userPhone && (
+            <div css={s.errorMsg}>{errorMsg.userPhone}</div>
+          )}
+        </div>
+        <div css={s.sort}>
+          <div css={s.myProfile}>사업자번호</div>
+          <div css={s.user}>
+            <input
+              type="text"
+              css={s.updateUserDetail}
+              value={user.userBusinessNumber}
+              readOnly
+            />
+          </div>
+        </div>
+        <div css={s.sort}>
+          <div css={s.myProfile_market}>마케팅 수신동의</div>
+          {user.marketingAgreed === true ? (
+            <input
+              css={s.checkBoxInput}
+              type="checkbox"
+              checked
+              onClick={handleCheckBox}
+            />
+          ) : (
+            <input
+              css={s.checkBoxInput}
+              type="checkbox"
+              onClick={handleCheckBox}
+            />
+          )}
+          <div css={s.myProfile_market_div}>
+          무료배송, 할인쿠폰 등 혜택/정보 수신 동의(선택)
+          </div>
+        </div>
 
-  &:hover{
-    background-color: #f0f0f0;
-  }
-
-`
-export const updateConfirm = css`
-  border: 1px solid #808;
-  border-radius: 3px;
-  color: #808;
-  background-color: #fff;
-  margin: 0px 10px;
-  padding: 10px 20px;
-  display: inline-block;
-  text-align: center;
-  cursor: pointer;
-`
-
-export const updateUserDetail = css`
-  border: none;
-  font-size: 19px;
-  width: 100%;
-  padding-left: 20px;
-  &:read-only {
-    background-color: #f0f0f0;
-    color: #333;
-    cursor: default;
-  }
-  
-`
-
-export const errorMsg = css`
-  color: red;
-  font-size: 12px;
-  margin-top: 5px;
-  margin-left: 270px;
-
-  
-`
-
-export const successMsg = css`
-  color: green;
-  font-size: 12px;
-  margin-top: 5px;
-  margin-left: 270px;
-`
+        <div css={s.buttonContainer}>
+          <button type="submit" css={s.updateConfirm} onClick={handlePutUser}>저장하기</button>
+          <button css={s.updateConfirm} onClick={() => { navigate(-1); }}>
+            취소하기
+            
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+}
